Ignore RTN_UNREACHABLE routes in route_dst_netlink.
According to rtnetlink(7), such routes are "an unreachable destination."
I get such a route when I ifdown my he-ipv6 interface:

$ /sbin/route -n -A inet6
Kernel IPv6 routing table
Destination                    Next Hop                   Flag Met Ref Use If
::/0                           ::                         !n   -1  1 26122 lo

The problem with not ignoring such routes is that Nmap will think that
the interface to use is lo, and consequently that all the targets are
localhost addresses. Ping scan will succeed with a localhost-response,
but trying to send any packets will fail with "destination unreachable."

Maybe we should do the same thing for these additional values of
rtm_type?

              RTN_BLACKHOLE     a packet dropping route
              RTN_PROHIBIT      a packet rejection route

git-svn-id: https://svn.nmap.org/nmap@29312 e0a8ed71-7df4-0310-8962-fdc924857419